test: refactor to framework
