+++ app/components/UI/Bridge/_mocks_/initialState.ts
@@ -378,6 +378,9 @@ export const initialState = {
               scopes: [EthScope.Eoa],
               metadata: {
                 lastSelected: 0,
+                keyring: {
+                  type: 'HD Key Tree',
+                },
               },
             },
             [solanaAccountId]: {
